requirements.txt
Django==4.0.4
mysqlclient==2.1.0
django-import-export==2.8.0
gunicorn==20.1.0
